+++ hw07-atm2/src/main/java/ru/otus/vsh/hw07/api/AtmBuilder.java
@@ -2,6 +2,8 @@ package ru.otus.vsh.hw07.api;
 
 import javax.annotation.Nonnull;
 import java.util.Map;
+
+import ru.otus.vsh.hw07.api.listeners.AtmValueChangeListener;
 import ru.otus.vsh.hw07.impl.AtmImpl;
 
 public class AtmBuilder {
