+++ hw07-atm2/src/main/java/ru/otus/vsh/hw07/api/listeners/AtmInitiateListener.java
@@ -1,4 +1,4 @@
-package ru.otus.vsh.hw07.api;
+package ru.otus.vsh.hw07.api.listeners;
 
 public interface AtmInitiateListener {
     void onInitiate(String reason);
