+++ hw07-atm2/src/main/java/ru/otus/vsh/hw07/api/listeners/AtmValueChangeListener.java
@@ -1,4 +1,6 @@
-package ru.otus.vsh.hw07.api;
+package ru.otus.vsh.hw07.api.listeners;
+
+import ru.otus.vsh.hw07.api.ValueChangeOperation;
 
 public interface AtmValueChangeListener {
     void onValueChange(long delta, ValueChangeOperation operation);
